--- models/categorias.js
@@ -1,37 +1,0 @@
-'use strict';
-const { Model } = require('sequelize');
-
-module.exports = (sequelize, DataTypes) => {
-    class Categories extends Model {
-        /**
-        * Helper method for defining associations.
-        * This method is not a part of Sequelize lifecycle.
-        * The `models/index` file will call this method automatically.
-        */
-        static associate(models) {
-            // define association here
-        }
-    };
-    Categories.init({
-        id: {
-            type: DataTypes.INTEGER,
-            autoIncrement: true,
-            primaryKey: true,
-            unique: true
-        },
-        name: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-        description: {
-            type: DataTypes.STRING,
-            allowNull: false
-        },
-        deletedAt: DataTypes.DATE
-    }, {
-        sequelize,
-        modelName: 'Categories',
-    });
-    return Categories;
-};
-
