+++ models/categories.js
@@ -1,0 +1,25 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class Categories extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+    }
+  };
+  Categories.init({
+    name: DataTypes.STRING,
+    description: DataTypes.TEXT,
+    deleteAt: DataTypes.DATE
+  }, {
+    sequelize,
+    modelName: 'Categories',
+  });
+  return Categories;
+};
